+++ packages/js/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.2.6] - 2020-02-06
+### Added
+- New method `validateDeviceId()` to assure the `deviceId` is available after page refresh.
 ### Fixed
 - Make sure `getDevices()` methods always return devices with both deviceId and label even if the browser does not have camera or microphone permissions
 ### Changed